Bug:
The Predicate was return true when the test method was called even though the predicates were added using the and method

When I added predicates using the "and" method, the model I thought of was the Collection such as ArrayList where the
Objects act as container.  However, each time a Predicate  is added to the Predicate using the and method a new Predicate
is returned.